requirements.txt
flask==2.0.1
line-bot-sdk==2.0.0
python-dotenv==0.19.0
gunicorn==20.1.0 